add skip lock file flag

+++ internal/exec/terraform.go
@@ -2,11 +2,12 @@ package exec
 
 import (
 	"fmt"
-	"github.com/pkg/errors"
-	"github.com/spf13/cobra"
 	"os"
 	"path"
 	"strings"
+
+	"github.com/pkg/errors"
+	"github.com/spf13/cobra"
 
 	cfg "github.com/cloudposse/atmos/pkg/config"
 	"github.com/cloudposse/atmos/pkg/schema"
@@ -14,9 +15,10 @@ import (
 )
 
 const (
-	autoApproveFlag = "-auto-approve"
-	outFlag         = "-out"
-	varFileFlag     = "-var-file"
+	autoApproveFlag           = "-auto-approve"
+	outFlag                   = "-out"
+	varFileFlag               = "-var-file"
+	skipTerraformLockFileFlag = "--skip-lock-file"
 )
 
 // ExecuteTerraformCmd parses the provided arguments and flags and executes terraform commands
@@ -81,8 +83,10 @@ func ExecuteTerraform(info schema.ConfigAndStacksInfo) error {
 			u.LogWarning(cliConfig, err.Error())
 		}
 
-		u.LogTrace(cliConfig, "Deleting '.terraform.lock.hcl' file")
-		_ = os.Remove(path.Join(componentPath, ".terraform.lock.hcl"))
+		if !u.SliceContainsString(info.AdditionalArgsAndFlags, skipTerraformLockFileFlag) {
+			u.LogTrace(cliConfig, "Deleting '.terraform.lock.hcl' file")
+			_ = os.Remove(path.Join(componentPath, ".terraform.lock.hcl"))
+		}
 
 		u.LogTrace(cliConfig, fmt.Sprintf("Deleting terraform varfile: %s\n", varFile))
 		_ = os.Remove(path.Join(componentPath, varFile))
